Add files via upload
In this task we use decision trees and random forest with feature importance to predict which leads are more likely to convert into paying customers based on their behavior. The fundamental goal of the customers is to get the highest F1 score possible. In order to achieve that, we built tree-based models. and selected the most important features.